Rename things to Render rather than Color
