+++ patchwork-dispatcher/build.gradle
@@ -1,5 +1,5 @@
 archivesBaseName = "patchwork-dispatcher"
-version = getSubprojectVersion(project, "0.2.0")
+version = getSubprojectVersion(project, "0.3.0")
 
 dependencies {
 	compile project(path: ':patchwork-fml', configuration: 'dev')
