+++ patchwork-events-rendering/src/main/resources/assets/patchwork-events-rendering/icon.png
