+++ patchwork-events-rendering/src/main/resources/patchwork-events-rendering.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "net.patchworkmc.mixin.event.colors",
+  "package": "net.patchworkmc.mixin.event.render",
   "compatibilityLevel": "JAVA_8",
   "client": [
     "MixinItemColors",
